Making get_u work with absolute time

+++ bioptim/optimization/non_linear_program.py
@@ -382,4 +382,4 @@ class NonLinearProgram:
         """
         if node_idx < 0 or node_idx > self.ns:
             return ValueError(f"node_index out of range [0:{self.ns}]")
-        return self.tf / self.ns * node_idx
+        return float(self.tf / self.ns * node_idx)
